Split android and iOS tests
