+++ build.gradle
@@ -189,15 +189,11 @@ task xcuiTest( type: Test ) {
     testLogging.showStandardStreams = true
     testLogging.exceptionFormat = 'full'
     filter {
-        includeTestsMatching '*.appium.element.generation.ios.*'
-        includeTestsMatching '*.appium.AppiumFluentWaitTest'
         includeTestsMatching 'io.appium.java_client.ios.*'
         includeTestsMatching '*.pagefactory_tests.XCUITModeTest'
         includeTestsMatching '*.pagefactory_tests.widget.tests.combined.*'
         includeTestsMatching '*.pagefactory_tests.widget.tests.ios.*'
-        includeTestsMatching '*.StartingAppLocallyTest.startingIOSAppWithCapabilitiesAndServiceTest'
-        includeTestsMatching '*.StartingAppLocallyTest.startingIOSAppWithCapabilitiesAndFlagsOnServerSideTest'
-        exclude '**/UIAutomationTest.class'
+        includeTestsMatching 'io.appium.java_client.service.local.StartingAppLocallyIosTest'
         exclude '**/IOSScreenRecordTest.class'
         exclude '**/ImagesComparisonTest.class'
     }
@@ -213,7 +209,9 @@ task uiAutomationTest( type: Test ) {
         includeTestsMatching '*.AndroidAppStringsTest'
         includeTestsMatching '*.pagefactory_tests.widget.tests.android.*'
         includeTestsMatching '*.pagefactory_tests.widget.tests.AndroidPageObjectTest'
-        includeTestsMatching 'io.appium.java_client.service.local.*'
+        includeTestsMatching 'io.appium.java_client.service.local.StartingAppLocallyAndroidTest'
+        includeTestsMatching 'io.appium.java_client.service.local.ServerBuilderTest'
+        includeTestsMatching 'io.appium.java_client.service.local.ThreadSafetyTest'
     }
 }
